feat: implement useMarqueeOverflow hook for overflow detection and refactor MarqueeText component
